@@ -9,9 +9,9 @@ homepage = "https://github.com/noneplugin/nonebot-plugin-wordle"
 repository = "https://github.com/noneplugin/nonebot-plugin-wordle"
 
 [tool.poetry.dependencies]
-python = "^3.8"
-nonebot2 = "^2.2.0"
-nonebot-plugin-alconna = ">=0.38.0,<1.0.0"
+python = "^3.9"
+nonebot2 = "^2.3.0"
+nonebot-plugin-alconna = ">=0.45.1,<1.0.0"
 nonebot-plugin-session = "^0.3.0"
 pillow = "^10.0.0"
 pyspellchecker = "^0.8.0"
@@ -19,13 +19,13 @@ pyspellchecker = "^0.8.0"
 [tool.poetry.group.dev.dependencies]
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 pythonPlatform = "All"
 typeCheckingMode = "basic"
 
 [tool.ruff]
 line-length = 88
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
